+++ mypy.ini
@@ -19,7 +19,7 @@ ignore_missing_imports = True
 [mypy-gds_helpers]
 ignore_missing_imports = True
 
-[mypy-pyaups.*]
+[mypy-pyaustraliapost.*]
 ignore_missing_imports = True
 
 [mypy-pycanadapost.*]
