--- purplship/carriers/aups/shipping/__init__.py
@@ -1,8 +1,0 @@
-from purplship.carriers.aups.shipping.tracking_items import (
-    parse_track_items_response,
-    track_items_request,
-)
-from purplship.carriers.aups.shipping.shipping_price import (
-    shipping_price_request,
-    parse_shipping_price_response,
-)
