+++ purplship/carriers/australiapost/postage/calculate_international_parcel.py
@@ -1,9 +1,9 @@
 from purplship.core.models import RateRequest
 from purplship.core.units import WeightUnit, Weight
-from pyaups.international_letter_postage import ServiceRequest
+from pyaustraliapost.international_parcel_postage import ServiceRequest
 
 
-def calculate_international_letter_request(payload: RateRequest) -> ServiceRequest:
+def calculate_international_parcel_request(payload: RateRequest) -> ServiceRequest:
     weight_unit: WeightUnit = WeightUnit[payload.parcel.weight_unit or "KG"]
     request = ServiceRequest(
         country_code=payload.recipient.country_code,
