--- purplship/package/mappers/aups/__init__.py
@@ -1,3 +1,0 @@
-from purplship.package.mappers.aups.mapper import Mapper
-from purplship.package.mappers.aups.proxy import Proxy
-from purplship.package.mappers.aups.settings import Settings
