+++ purplship/package/mappers/australiapost/__init__.py
@@ -1,0 +1,3 @@
+from purplship.package.mappers.australiapost.mapper import Mapper
+from purplship.package.mappers.australiapost.proxy import Proxy
+from purplship.package.mappers.australiapost.settings import Settings
