+++ setup.package.py
@@ -24,7 +24,7 @@ setup(name='purplship.package',
             'py-soap',
             'py-ups',
             'py-canadapost',
-            'py-aups',
+            'py-australiapost',
             'py-sendle',
             'py-usps',
       ],
