+++ setup.py
@@ -14,6 +14,7 @@ setup(name='purplship',
       license='LGPLv3',
       packages=find_packages(".", exclude=["tests*"]),
       install_requires=[
+            'attrs',
             'jstruct',
             'xmltodict',
             'lxml',
@@ -23,7 +24,7 @@ setup(name='purplship',
             'py-soap',
             'py-ups',
             'py-canadapost',
-            'py-aups',
+            'py-australiapost',
             'py-sendle',
             'py-usps',
       ],
