--- tests/aups/logistic/__init__.py
@@ -1,2 +1,0 @@
-from tests.aups.logistic.tracking import *
-from tests.aups.logistic.rate import *
