+++ tests/australiapost/logistic/__init__.py
@@ -1,0 +1,2 @@
+from tests.australiapost.logistic.tracking import *
+from tests.australiapost.logistic.rate import *
