+++ tests/australiapost/logistic/fixture.py
@@ -1,5 +1,5 @@
 import purplship.package as api
 
-gateway = api.gateway["aups"].create(
+gateway = api.gateway["australiapost"].create(
     {"api_key": "username", "password": "password", "account_number": "1234567"}
 )
